Added the ability to rate games. Might just call it here given the time constraints and I've completed the requirements for the project. Probably best not to break it.

+++ ServerSide_GameRater/Shared/Rating.cs
@@ -11,6 +11,6 @@ namespace ServerSide_GameRater.Shared
         public int ratingID {  get; set; } = 0;
         public int gameID { get; set; } = 0;
         public int userID { get; set; } = 0;
-        public float rating { get; set; } = 0.0f;
+        public double rating { get; set; } = 0.0f;
     }
 }
